refactor(basecheckbox): refactor to extend BaseInput interface
