+++ packages/components/inputs/src/base-checkbox/BaseCheckbox.styles.ts
@@ -1,18 +1,19 @@
 import { css } from 'emotion';
 import tokens from '@contentful/f36-tokens';
-import { BaseCheckboxProps } from './BaseCheckbox';
+import type { BaseCheckboxProps } from './types';
 
 const getStyles = ({
   isDisabled,
   type,
 }: Pick<BaseCheckboxProps, 'isDisabled' | 'type'>) => ({
   wrapper: css({
+    alignItems: 'center',
     display: 'inline-flex',
     position: 'relative',
     margin: '0',
   }),
   input: css({
-    cursor: !isDisabled ? 'pointer' : 'initial',
+    cursor: isDisabled ? 'not-allowed' : 'pointer',
     height: tokens.spacingM,
     margin: 0,
     opacity: 0,
@@ -25,11 +26,6 @@ const getStyles = ({
         boxShadow: tokens.glowPrimary,
       },
     },
-  }),
-  label: css({
-    color: tokens.gray900,
-    fontSize: tokens.fontSizeM,
-    lineHeight: tokens.lineHeightM,
   }),
 });
